refactor(arto-markdown): confine the parser to an engine module
Code that used pulldown-cmark types was spread over the crate root and
five sibling modules, so an engine swap would touch every file. The
alert rewrite, the Mermaid and math event transforms, the source line
injection, the heading outline and the selection source map now live
under `engine/`, and the crate root only sees `engine::render` with an
engine-neutral result (HTML, headings, table lines) and
`extract_source_selection`. No parser type crosses that boundary, which
keeps the swap a change inside one directory.

Behaviour is unchanged; the modules moved without edits beyond their
imports.
